refactor: email client stub
