+++ packages/backend/src/email/client-stub.ts
@@ -1,14 +1,12 @@
-import { compileEmailTemplate } from '../email/helpers'
+import { compileEmailTemplate } from './helpers'
 
 type SentMessage = {
   to: string
   subject: string
-  text?: string
-  html?: string
+  html: string
 }
 
 export interface EmailClientStub {
-  sendMail: (to: string, subject: string, text: string) => Promise<void>
   sendMailFromTemplate: (
     to: string,
     subject: string,
@@ -21,10 +19,6 @@ export interface EmailClientStub {
 export function createEmailClientStub(): EmailClientStub {
   const sentMessages: SentMessage[] = []
   return {
-    async sendMail(to: string, subject: string, text: string) {
-      sentMessages.push({ to, subject, text })
-      return Promise.resolve()
-    },
     async sendMailFromTemplate(
       to: string,
       subject: string,
