+++ packages/backend/src/email/client.ts
@@ -3,11 +3,6 @@ import nodemailer, { SentMessageInfo, Transporter } from 'nodemailer'
 import { compileEmailTemplate } from './helpers'
 
 export interface EmailClient {
-  sendMail: (
-    to: string,
-    subject: string,
-    text: string,
-  ) => Promise<SentMessageInfo>
   sendMailFromTemplate: (
     to: string,
     subject: string,
@@ -28,22 +23,8 @@ interface SmtpConfig {
 export function createEmailClient(config: SmtpConfig): EmailClient {
   const transporter = nodemailer.createTransport(config)
   return {
-    sendMail: partial(sendMail, transporter),
     sendMailFromTemplate: partial(sendMailFromTemplate, transporter),
   }
-}
-
-function sendMail(
-  transporter: Transporter,
-  to: string,
-  subject: string,
-  text: string,
-): Promise<SentMessageInfo> {
-  return transporter.sendMail({
-    to,
-    subject,
-    text,
-  })
 }
 
 async function sendMailFromTemplate(
